move files and adapt to mobiles
