+++ package.json
@@ -80,7 +80,7 @@
     "@deck.gl/mesh-layers": "^8.5.4",
     "@konturio/default-icons": "^2.2.0",
     "@konturio/default-theme": "3.1.1",
-    "@konturio/ui-kit": "^3.5.2",
+    "@konturio/ui-kit": "^3.5.3",
     "@loaders.gl/core": "^3.2.3",
     "@nebula.gl/edit-modes": "1.0.3",
     "@nebula.gl/layers": "1.0.3",
