--- src/features/layers_panel/components/MapLayersPanel/MapLayersPanel.module.css
@@ -1,74 +1,0 @@
-.panelContainer {
-  display: flex;
-  flex-direction: column;
-  overflow: auto;
-  box-shadow: var(--elevation-1);
-}
-.panel {
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-  z-index: 1;
-
-  & > div {
-    overflow: hidden;
-  }
-
-  &.collapse {
-    height: auto;
-    & div:not(:first-child) {
-      display: none;
-    }
-  }
-}
-
-.scrollable {
-  overflow-y: auto;
-  height: 100%;
-}
-
-.panelIcon {
-  display: none;
-}
-/* desktop only */
-@media screen and (min-width: 1530px) {
-  .panelContainer.open {
-    flex: 1 2 auto;
-  }
-}
-
-/* laptop and below */
-@media screen and (max-width: 1530px) {
-  .panelContainer {
-    overflow: unset;
-    box-shadow: none;
-  }
-}
-
-/* mobile */
-@media screen and (max-width: 960px) {
-  .panel.show {
-    position: absolute;
-    margin: 0;
-    width: 100vw;
-    height: calc(100% - var(--app-header-height));
-    align-items: center;
-    & > div {
-      width: 100%;
-    }
-  }
-  .panel.collapse {
-    display: none;
-  }
-
-  .scrollable {
-    width: 100%;
-  }
-
-  .panelIcon {
-    &.show {
-      visibility: visible;
-      display: flex;
-    }
-  }
-}
