+++ src/features/layers_panel/components/PanelContent/PanelContent.module.css
@@ -1,0 +1,11 @@
+.scrollable {
+  overflow-y: auto;
+  height: 100%;
+}
+
+/* mobile */
+@media screen and (max-width: 960px) {
+  .scrollable {
+    width: 100%;
+  }
+}
