+++ src/features/layers_panel/components/PanelContent/PanelContent.tsx
@@ -1,0 +1,10 @@
+import { LayersTree } from '../LayersTree/LayersTree';
+import s from './PanelContent.module.css';
+
+export function PanelContent() {
+  return (
+    <div className={s.scrollable}>
+      <LayersTree />
+    </div>
+  );
+}
