+++ src/features/layers_panel/index.tsx
@@ -1,15 +1,1 @@
-import { MapLayerPanel } from './components';
-import { LayersTree } from './components/LayersTree/LayersTree';
-import s from './components/MapLayersPanel/MapLayersPanel.module.css';
-
-export function MapLayersList() {
-  return <MapLayerPanel />;
-}
-
-export function LayersPanelContent() {
-  return (
-    <div className={s.scrollable}>
-      <LayersTree />
-    </div>
-  );
-}
+export { PanelContent as LayersPanelContent } from './components';
