--- src/features/legend_panel/components/LegendPanel/LegendPanel.module.css
@@ -1,85 +1,0 @@
-.panelContainer {
-  display: flex;
-  flex-direction: column;
-  box-shadow: var(--elevation-1);
-  &.open {
-    overflow: auto;
-  }
-}
-.legendPanel {
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-
-  &.collapse {
-    height: auto;
-    & div:not(:first-child) {
-      display: none;
-    }
-  }
-}
-.content {
-  height: auto;
-}
-.scrollable {
-  overflow-y: auto;
-  overflow-x: hidden;
-  padding: var(--double-unit);
-}
-
-.panelIcon {
-  display: none;
-}
-
-.legendContainer {
-  margin-bottom: var(--unit);
-  & .layerControl {
-    padding: 0;
-    &:hover {
-      background: none;
-    }
-  }
-}
-
-.legendLabel {
-  display: flex;
-  flex-flow: row;
-  gap: var(--half-unit);
-  margin-bottom: var(--half-unit);
-}
-
-.legendBody {
-  margin-left: var(--double-unit);
-}
-
-.alignRight {
-  margin-left: auto;
-}
-
-/* mobile */
-@media screen and (max-width: 960px) {
-  .legendPanel.show {
-    position: absolute;
-    margin: 0;
-    width: 100vw;
-    height: calc(100% - var(--app-header-height));
-    align-items: center;
-    & > div {
-      width: 100%;
-    }
-  }
-  .legendPanel.collapse {
-    display: none;
-  }
-
-  .scrollable {
-    width: 100%;
-  }
-
-  .panelIcon {
-    &.show {
-      visibility: visible;
-      display: flex;
-    }
-  }
-}
